requirements.txt
tweepy==3.9.0
requests-cache==0.5.2
pandas==1.1.0
python-dotenv==0.14.0